requirements.txt
playwright
beautifulsoup4
lxml
httpx
python-dotenv
fake-useragent
tenacity
structlog
rich
